+++ ts/tests/vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    dir: "./test",
+    testTimeout: 60000,
+    hookTimeout: 30000,
+  },
+});
